Fixes to handle block results better, allow persistence of peer state

+++ src/main/java/convex/net/Message.java
@@ -81,7 +81,7 @@ public class Message {
 			case TRANSACT: return (Long) ((AVector<?>)payload).get(0);
 			
 			// Result is a special record type
-			case RESULT: return ((Result)payload).getID(); 
+			case RESULT: return (Long)((Result)payload).getID(); 
 			
 			default: return null;
 		}
